+++ play.py
@@ -63,6 +63,8 @@ class WechatAutoJump(object):
             os.system('adb pull /sdcard/1.png state.png')
         elif self.phone == 'IOS':
             self.client.screenshot('state.png')
+        if not os.path.exists('state.png'):
+            raise NameError('Cannot obtain screenshot from your phone! Please follow the instructions in readme!')
 
         if self.debug:
             shutil.copyfile('state.png', os.path.join(self.debug, 'state_{:03d}.png'.format(self.step)))
@@ -97,7 +99,7 @@ class WechatAutoJump(object):
         state_cut = state[sym_tl[0]:sym_br[0], sym_tl[1]:sym_br[1]]
         target_pos = None
         for target in self.jump_file:
-            pos = multi_scale_search(target, state_cut, 0.4, 15)
+            pos = multi_scale_search(target, state_cut, 0.3, 15)
             if target_pos is None or pos[-1] > target_pos[-1]:
                 target_pos = pos
         return np.array([(target_pos[0]+target_pos[2])//2, (target_pos[1]+target_pos[3])//2]) + sym_tl
@@ -122,6 +124,8 @@ class WechatAutoJump(object):
         press_time = distance * self.sensitivity
         press_time = int(np.rint(press_time))
         press_h, press_w = int(0.82*self.resolution[0]), self.resolution[1]//2
+        press_h += np.random.randint(-50,50)
+        press_w += np.random.randint(-50,50)
         if self.phone == 'Android':
             cmd = 'adb shell input swipe {} {} {} {} {}'.format(press_w, press_h, press_w, press_h, press_time)
             print(cmd)
@@ -152,7 +156,7 @@ class WechatAutoJump(object):
             self.debugging()
         self.jump(self.player_pos, self.target_pos)
         self.step += 1
-        time.sleep(1.5)
+        time.sleep(1.4 + np.random.random())
 
     def run(self):
         try:
